+++ plugins/pvio/pvio_socket.c
@@ -66,10 +66,10 @@
 /* Function prototypes */
 my_bool pvio_socket_set_timeout(MARIADB_PVIO *pvio, enum enum_pvio_timeout type, int timeout);
 int pvio_socket_get_timeout(MARIADB_PVIO *pvio, enum enum_pvio_timeout type);
-size_t pvio_socket_read(MARIADB_PVIO *pvio, uchar *buffer, size_t length);
-size_t pvio_socket_async_read(MARIADB_PVIO *pvio, uchar *buffer, size_t length);
-size_t pvio_socket_async_write(MARIADB_PVIO *pvio, const uchar *buffer, size_t length);
-size_t pvio_socket_write(MARIADB_PVIO *pvio, const uchar *buffer, size_t length);
+ssize_t pvio_socket_read(MARIADB_PVIO *pvio, uchar *buffer, size_t length);
+ssize_t pvio_socket_async_read(MARIADB_PVIO *pvio, uchar *buffer, size_t length);
+ssize_t pvio_socket_async_write(MARIADB_PVIO *pvio, const uchar *buffer, size_t length);
+ssize_t pvio_socket_write(MARIADB_PVIO *pvio, const uchar *buffer, size_t length);
 int pvio_socket_wait_io_or_timeout(MARIADB_PVIO *pvio, my_bool is_read, int timeout);
 my_bool pvio_socket_blocking(MARIADB_PVIO *pvio, my_bool value, my_bool *old_value);
 my_bool pvio_socket_connect(MARIADB_PVIO *pvio, MA_PVIO_CINFO *cinfo);
@@ -262,7 +262,7 @@ int pvio_socket_get_timeout(MARIADB_PVIO *pvio, enum enum_pvio_timeout type)
      -1              on error
                      
 */   
-size_t pvio_socket_read(MARIADB_PVIO *pvio, uchar *buffer, size_t length)
+ssize_t pvio_socket_read(MARIADB_PVIO *pvio, uchar *buffer, size_t length)
 {
   ssize_t r= -1;
 #ifndef _WIN32
@@ -298,7 +298,7 @@ size_t pvio_socket_read(MARIADB_PVIO *pvio, uchar *buffer, size_t length)
       errno= WSAGetLastError();
       return -1;
     }
-    r= dwBytes;
+    r= (ssize_t)dwBytes;
   }
 #endif
   return r;
@@ -325,7 +325,7 @@ size_t pvio_socket_read(MARIADB_PVIO *pvio, uchar *buffer, size_t length)
      -1              on error
                      
 */   
-size_t pvio_socket_async_read(MARIADB_PVIO *pvio, uchar *buffer, size_t length)
+ssize_t pvio_socket_async_read(MARIADB_PVIO *pvio, uchar *buffer, size_t length)
 {
   ssize_t r= -1;
 #ifndef _WIN32
@@ -387,7 +387,7 @@ ssize_t ma_send(int socket, const uchar *buffer, size_t length, int flags)
      -1              on error
                      
 */   
-size_t pvio_socket_async_write(MARIADB_PVIO *pvio, const uchar *buffer, size_t length)
+ssize_t pvio_socket_async_write(MARIADB_PVIO *pvio, const uchar *buffer, size_t length)
 {
   ssize_t r= -1;
   struct st_pvio_socket *csock= NULL;
@@ -436,7 +436,7 @@ size_t pvio_socket_async_write(MARIADB_PVIO *pvio, const uchar *buffer, size_t l
      -1              on error
                      
 */
-size_t pvio_socket_write(MARIADB_PVIO *pvio, const uchar *buffer, size_t length)
+ssize_t pvio_socket_write(MARIADB_PVIO *pvio, const uchar *buffer, size_t length)
 {
   ssize_t r= -1;
   struct st_pvio_socket *csock= NULL;
